@cmsbuild 2015-07-30
This pull request is fully signed and it will be integrated in one of the next CMSSW_7_6_X IBs  or unless it breaks tests. This pull request requires discussion in the ORP meeting before it's merged. @davidlange6, @degano, @smuzaffar